Prompt for tool check every time.
The previous behavior was to only prompt the user to `go get`
the `gogetdoc` tool once.  If they elected not to, future
attempts to get documentation would just hang.

Instead, we prompt the user go `go get` the tool every time
they attempt to get documentation (when the tool can't be found).
This way the user can decide to postpone installing the tool.